feat: keep it DRY
