+++ packages/hdwallet-phantom/src/solana.ts
@@ -1,12 +1,6 @@
 import * as core from "@shapeshiftoss/hdwallet-core";
 import {
-  AddressLookupTableAccount,
-  ComputeBudgetProgram,
   PublicKey,
-  SystemProgram,
-  TransactionInstruction,
-  TransactionMessage,
-  VersionedTransaction,
 } from "@solana/web3.js";
 
 import { PhantomSolanaProvider } from "./types";
@@ -15,61 +9,13 @@ export type SolanaAccount = {
   publicKey: PublicKey;
 };
 
-function toTransactionInstructions(instructions: core.SolanaTxInstruction[]): TransactionInstruction[] {
-  return instructions.map(
-    (instruction) =>
-      new TransactionInstruction({
-        keys: instruction.keys.map((key) => Object.assign(key, { pubkey: new PublicKey(key.pubkey) })),
-        programId: new PublicKey(instruction.programId),
-        data: instruction.data,
-      })
-  );
-}
-
-function buildTransaction(msg: core.SolanaSignTx, address: string): VersionedTransaction {
-  const instructions = toTransactionInstructions(msg.instructions ?? []);
-
-  const value = Number(msg.value);
-  if (!isNaN(value) && value > 0 && msg.to) {
-    instructions.push(
-      SystemProgram.transfer({
-        fromPubkey: new PublicKey(address),
-        toPubkey: new PublicKey(msg.to),
-        lamports: value,
-      })
-    );
-  }
-
-  if (msg.computeUnitLimit !== undefined) {
-    instructions.push(ComputeBudgetProgram.setComputeUnitLimit({ units: msg.computeUnitLimit }));
-  }
-
-  if (msg.computeUnitPrice !== undefined) {
-    instructions.push(ComputeBudgetProgram.setComputeUnitPrice({ microLamports: msg.computeUnitPrice }));
-  }
-
-  const addressLookupTableAccounts = msg.addressLookupTableAccountInfos?.map((accountInfo) => {
-    return new AddressLookupTableAccount({
-      key: new PublicKey(accountInfo.key),
-      state: AddressLookupTableAccount.deserialize(new Uint8Array(accountInfo.data)),
-    });
-  });
-
-  const message = new TransactionMessage({
-    payerKey: new PublicKey(address),
-    instructions,
-    recentBlockhash: msg.blockHash,
-  }).compileToV0Message(addressLookupTableAccounts);
-
-  return new VersionedTransaction(message);
-}
 
 export async function solanaSignTx(
   msg: core.SolanaSignTx,
   provider: PhantomSolanaProvider,
   address: string
 ): Promise<core.SolanaSignedTx | null> {
-  const transaction = buildTransaction(msg, address);
+  const transaction = core.buildTransaction(msg, address);
   const signedTransaction = await provider.signTransaction(transaction);
   return {
     serialized: Buffer.from(signedTransaction.serialize()).toString("base64"),
@@ -82,7 +28,7 @@ export async function solanaSendTx(
   provider: PhantomSolanaProvider,
   address: string
 ): Promise<core.SolanaTxSignature | null> {
-  const transaction = buildTransaction(msg, address);
+  const transaction = core.buildTransaction(msg, address);
   const { signature } = await provider.signAndSendTransaction(transaction);
   return { signature };
 }
